+++ pandas/core/reshape/api.py
@@ -1,7 +1,7 @@
 # flake8: noqa
 
 from pandas.core.reshape.concat import concat
-from pandas.core.reshape.reshape import melt
+from pandas.core.reshape.melt import melt
 from pandas.core.reshape.merge import (
     merge, ordered_merge, merge_ordered, merge_asof)
 from pandas.core.reshape.pivot import pivot_table, crosstab
